requirements.txt
# --------- pytorch --------- #
pytorch-lightning>=1.4.1

# --------- hydra --------- #
hydra-core==1.1.0.rc1
hydra-colorlog==1.1.0.dev1
hydra-optuna-sweeper==1.1.0.dev2
# hydra-ax-sweeper==1.1.0
# hydra-ray-launcher==0.1.2
# hydra-submitit-launcher==1.1.0

# --------- loggers --------- #
wandb>=0.10.30
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit
black
isort
flake8

# --------- others --------- #
jupyterlab
python-dotenv
rich
pytest
sh
scikit-learn
seaborn
pudb
# dvc


# --------- kuielab --------- #
coloredlogs
loguru
musdb
museval
wget